requirements.txt
flask
mesa
numpy
solara
matplotlib
requests
google-generativeai
